requirements.txt.txt
# Name                    Version                   Build  Channel
_tflow_select             2.1.0                       gpu
absl-py                   0.9.0                    py37_0
asn1crypto                1.3.0                    py37_0
astor                     0.8.0                    py37_0
blas                      1.0                         mkl
blinker                   1.4                      py37_0
ca-certificates           2020.1.1                      0
cachetools                3.1.1                      py_0
certifi                   2020.4.5.1               py37_0
cffi                      1.14.0           py37h7a1dbc1_0
chardet                   3.0.4                 py37_1003
click                     7.1.1                      py_0
cloudpickle               1.3.0                    pypi_0    pypi
cryptography              2.8              py37h7a1dbc1_0
cudatoolkit               10.1.243             h74a9793_0
cudnn                     7.6.5                cuda10.1_0
cycler                    0.10.0                   pypi_0    pypi
decorator                 4.4.2                    pypi_0    pypi
gast                      0.2.2                    py37_0
google-auth               1.13.1                     py_0
google-auth-oauthlib      0.4.1                      py_2
google-pasta              0.2.0                      py_0
grpcio                    1.27.2           py37h351948d_0
h5py                      2.10.0           py37h5e291fa_0
hdf5                      1.10.4               h7ebc959_0
icc_rt                    2019.0.0             h0cc432a_1
idna                      2.9                        py_1
imageio                   2.8.0                    pypi_0    pypi
intel-openmp              2020.0                      166
joblib                    0.14.1                   pypi_0    pypi
keras                     2.3.1                         0
keras-applications        1.0.8                      py_0
keras-base                2.3.1                    py37_0
keras-preprocessing       1.1.0                      py_1
kiwisolver                1.2.0                    pypi_0    pypi
libprotobuf               3.11.4               h7bd577a_0
markdown                  3.1.1                    py37_0
matplotlib                3.2.1                    pypi_0    pypi
mkl                       2020.0                      166
mkl-service               2.3.0            py37hb782905_0
mkl_fft                   1.0.15           py37h14836fe_0
mkl_random                1.1.0            py37h675688f_0
networkx                  2.4                      pypi_0    pypi
numpy                     1.18.1           py37h93ca92e_0
numpy-base                1.18.1           py37hc3f5095_1
oauthlib                  3.1.0                      py_0
opencv-python             4.2.0.34                 pypi_0    pypi
openssl                   1.1.1f               he774522_0
opt_einsum                3.1.0                      py_0
pandas                    1.0.3            py37h47e9c7a_0
pillow                    7.1.1                    pypi_0    pypi
pip                       20.0.2                   py37_1
progressbar2              3.50.1                   pypi_0    pypi
protobuf                  3.11.4           py37h33f27b4_0
pyasn1                    0.4.8                      py_0
pyasn1-modules            0.2.7                      py_0
pycparser                 2.20                       py_0
pyjwt                     1.7.1                    py37_0
pyopenssl                 19.1.0                   py37_0
pyparsing                 2.4.7                    pypi_0    pypi
pyreadline                2.1                      py37_1
pysocks                   1.7.1                    py37_0
python                    3.7.7           h60c2a47_0_cpython
python-dateutil           2.8.1                      py_0
python-utils              2.4.0                    pypi_0    pypi
pytz                      2019.3                     py_0
pywavelets                1.1.1                    pypi_0    pypi
pyyaml                    5.3.1            py37he774522_0
requests                  2.23.0                   py37_0
requests-oauthlib         1.3.0                      py_0
rsa                       4.0                        py_0
scikit-image              0.16.2                   pypi_0    pypi
scikit-learn              0.22.2.post1             pypi_0    pypi
scipy                     1.4.1            py37h9439919_0
setuptools                46.1.3                   py37_0
six                       1.14.0                   py37_0
sqlite                    3.31.1               he774522_0
tensorboard               2.1.0                     py3_0
tensorflow                2.1.0           gpu_py37h7db9008_0
tensorflow-base           2.1.0           gpu_py37h55f5790_0
tensorflow-estimator      2.1.0              pyhd54b08b_0
tensorflow-gpu            2.1.0                h0d30ee6_0
tensorlayer               2.2.1                    pypi_0    pypi
termcolor                 1.1.0                    py37_1
urllib3                   1.25.8                   py37_0
vc                        14.1                 h0510ff6_4
vs2015_runtime            14.16.27012          hf0eaf9b_1
werkzeug                  0.14.1                   py37_0
wheel                     0.34.2                   py37_0
win_inet_pton             1.1.0                    py37_0
wincertstore              0.2                      py37_0
wrapt                     1.12.1           py37he774522_1
yaml                      0.1.7                hc54c509_2
zlib                      1.2.11               h62dcd97_3
